Move find-permit.js to processors
This will be getting more use so doesn't make sense to keep it off with the summary pages.
